+++ CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.0] - 2021-02-17
 ### Added
 - Add support for the `in_app_include` and `in_app_exclude` options
 - Add all log messages as breadcrumbs
 - Add support for Drush errors
 - Add default excluded exceptions & tags
+- Add PHP 8 support
 
 ### Fixed
 - Replace recommended guzzlehttp/psr7 with php-http/guzzle6-adapter